divided one reposirory to others, some changes in blocs
