+++ lib/src/common/network/repositories/interface_category_repository.dart
@@ -1,0 +1,5 @@
+import 'package:coffee_shop/src/features/menu/data/category_data.dart';
+
+abstract interface class ICategoryRepository {
+  Future<List<CategoryData>> loadOnlyCategories();
+}
