+++ lib/src/common/network/repositories/interface_order_repository.dart
@@ -1,0 +1,3 @@
+abstract interface class IOrderRepository {
+  Future<bool> sendOrder({required Map<String, int> products});
+}
