+++ lib/src/features/database/database_data_source.dart
@@ -2,7 +2,6 @@ import 'package:coffee_shop/src/features/database/coffee_database.dart';
 import 'package:coffee_shop/src/features/database/interface_savable_data_source.dart';
 import 'package:coffee_shop/src/features/menu/data/category_data.dart';
 import 'package:coffee_shop/src/features/menu/data/product_data.dart';
-import 'package:drift/drift.dart';
 import 'package:flutter/material.dart';
 
 final class DataBaseSource implements ISavableDataSource {
@@ -21,14 +20,14 @@ final class DataBaseSource implements ISavableDataSource {
 
   @override
   Future<List<CategoryData>> fetchOnlyCategories() async {
-    await database.into(database.categories).insert(CategoriesCompanion.insert(
-          id: const Value(2),
-          name: 'Abobab',
-        ));
     List<Category> allItems = await database.select(database.categories).get();
-
+    List<CategoryData> categories = [];
     debugPrint('items in database: $allItems');
-    throw UnimplementedError();
+    for (var category in allItems){
+      categories.add(CategoryData(id: category.id, name: category.name, products: []));
+    }
+    debugPrint('Categories: $categories');
+    return categories;
   }
 
   @override
